+++ Source/Core/Core/Core.vcxproj.filters
@@ -509,10 +509,7 @@
     <ClCompile Include="HW\SystemTimers.cpp">
       <Filter>HW %28Flipper/Hollywood%29</Filter>
     </ClCompile>
-    <ClCompile Include="DSP\assemble.cpp">
-      <Filter>DSPCore</Filter>
-    </ClCompile>
-    <ClCompile Include="DSP\disassemble.cpp">
+    <ClCompile Include="DSP\DSPAssembler.cpp">
       <Filter>DSPCore</Filter>
     </ClCompile>
     <ClCompile Include="DSP\DSPAccelerator.cpp">
@@ -525,6 +522,9 @@
       <Filter>DSPCore</Filter>
     </ClCompile>
     <ClCompile Include="DSP\DSPCore.cpp">
+      <Filter>DSPCore</Filter>
+    </ClCompile>
+    <ClCompile Include="DSP\DSPDisassembler.cpp">
       <Filter>DSPCore</Filter>
     </ClCompile>
     <ClCompile Include="DSPEmulator.cpp">
@@ -1038,10 +1038,7 @@
     <ClInclude Include="HW\SystemTimers.h">
       <Filter>HW %28Flipper/Hollywood%29</Filter>
     </ClInclude>
-    <ClInclude Include="DSP\assemble.h">
-      <Filter>DSPCore</Filter>
-    </ClInclude>
-    <ClInclude Include="DSP\disassemble.h">
+    <ClInclude Include="DSP\DSPAssembler.h">
       <Filter>DSPCore</Filter>
     </ClInclude>
     <ClInclude Include="DSP\DSPAccelerator.h">
@@ -1060,6 +1057,9 @@
       <Filter>DSPCore</Filter>
     </ClInclude>
     <ClInclude Include="DSP\DSPCore.h">
+      <Filter>DSPCore</Filter>
+    </ClInclude>
+    <ClInclude Include="DSP\DSPDisassembler.h">
       <Filter>DSPCore</Filter>
     </ClInclude>
     <ClInclude Include="DSPEmulator.h">
